requirements.txt
google-cloud-aiplatform>=0.3.0
pydantic>=2.0.0
PyQt6>=6.4.0
numpy>=1.24.0
soundfile>=0.12.1
librosa>=0.10.1
pytest>=7.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
xlsxwriter>=3.1.0
pyinstaller>=6.3.0
# Audio processing
SoundFile>=0.12.1
# UI
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0
# Testing
pytest-cov>=4.1.0
pytest-qt>=4.2.0 